LPS-54269 Move test logic from DBStoreTest into BaseStoreTestCase, this includes the same test AdvancedFileSystemStoreTest has
